// src/app/dashboard/kegpembelajaran/models/quiz.model.ts
/* eslint-disable eol-last */
export interface Quiz {
    question: string;
    options: string[];
    answer: string;
    explanation: string;
}
export interface Evaluasi {
    question: string;
    options: string[];
    answer: string;
}
  export interface Dua {
    question: string;
    options: string[];
    correctAnswers: string[];
  }